[#21][refactor]Log, LogType을 log 패키지로 이동
