+++ src/main/java/dev/idion/bladitodo/domain/board/Board.java
@@ -1,8 +1,8 @@
 package dev.idion.bladitodo.domain.board;
 
 import dev.idion.bladitodo.domain.entity.BaseEntity;
-import dev.idion.bladitodo.domain.entity.Log;
 import dev.idion.bladitodo.domain.list.List;
+import dev.idion.bladitodo.domain.log.Log;
 import java.util.ArrayList;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
